♻️ Move Open to internal/wrap and define Open function again
