+++ internal/wrap/open_test.go
@@ -1,10 +1,12 @@
-package sqlslog
+package wrap
 
 import (
 	"bytes"
 	"context"
 	"log/slog"
 	"testing"
+
+	"github.com/akm/sql-slog/public"
 )
 
 func TestOpen(t *testing.T) {
@@ -13,7 +15,7 @@ func TestOpen(t *testing.T) {
 	buf := bytes.NewBuffer(nil)
 	logger := slog.New(slog.NewJSONHandler(buf, &slog.HandlerOptions{Level: slog.LevelDebug}))
 
-	db, err := Open(ctx, "invalid-driver", "", Logger(logger))
+	db, err := Open(ctx, "invalid-driver", "", public.Logger(logger))
 	if err == nil {
 		t.Fatal("Expected error")
 	}
